+++ CHANGELOG.md
@@ -18,8 +18,10 @@ a release.
 ---
 
 ## [Unreleased]
+
+## [3.8.0] - 2022-07-17
 #### Added
-- Sluggable: Add support for DateTimeImmutable fields
+- Sluggable: Add support for `DateTimeImmutable` fields
 - Tree: [NestedSet] `childrenQueryBuilder()` to allow specifying sort order separately for each field
 - Tree: [NestedSet] Added option to reorder only direct children in `reorder()` method